ITK: update to ITK 3.18 with OTB patches

+++ Utilities/ITK/Utilities/itktiff/tif_luv.c
@@ -1,4 +1,4 @@
-/* $Id: tif_luv.c,v 1.2 2007-01-04 16:46:32 ibanez Exp $ */
+/* $Id: tif_luv.c,v 1.3 2009-11-15 21:04:22 ibanez Exp $ */
 
 /*
  * Copyright (c) 1997 Greg Ward Larson
@@ -875,7 +875,7 @@ LogLuv24toXYZ(uint32 p, float XYZ[3])
           /* decode luminance */
   L = LogL10toY(p>>14 & 0x3ff);
   if (L <= 0.) {
-    XYZ[0] = XYZ[1] = XYZ[2] = 0.;
+    XYZ[0] = XYZ[1] = XYZ[2] = (float)(0.0);
     return;
   }
           /* decode color */
@@ -1015,7 +1015,7 @@ LogLuv32toXYZ(uint32 p, float XYZ[3])
           /* decode luminance */
   L = LogL16toY((int)p >> 16);
   if (L <= 0.) {
-    XYZ[0] = XYZ[1] = XYZ[2] = 0.;
+    XYZ[0] = XYZ[1] = XYZ[2] = (float)(0.0);
     return;
   }
           /* decode color */
